pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "privault"
version = "0.1.0"
description = "Local-only AES-256-GCM encrypted personal vault. Zero cloud, zero telemetry, AI-resistant."
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["security", "privacy", "vault", "password-manager", "encryption"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "typer>=0.9.0",
    "cryptography>=41.0.0",
    "argon2-cffi>=23.1.0",
    "pyperclip>=1.8.2",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "types-pyperclip",
]

[project.scripts]
privault = "privault.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/privault"]

[tool.mypy]
strict = true
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/privault --cov-report=term-missing --cov-fail-under=85"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
